add Clock time correction via localTime

+++ src/main/java/cz/cvut/fel/pjv/MainApp.java
@@ -1,6 +1,5 @@
 package cz.cvut.fel.pjv;
 
-import cz.cvut.fel.pjv.PGN.PgnRefactor;
 import cz.cvut.fel.pjv.TimeClock.TheClock;
 import cz.cvut.fel.pjv.model.Game;
 import cz.cvut.fel.pjv.model.Player.ComputerPlayer;
@@ -133,7 +132,7 @@ public class MainApp extends Application {
         }
 
         // get current time value
-        int timeLefts;
+        double timeLefts;
         if (game.getGameRound() % 2 == 1) {
             timeLefts = game.getTimeLefts()[0];
         } else {
